requirements.txt
# ============================================================================
# Production Dependencies for Aquaculture ML Platform
# ============================================================================
#
# This file contains all production dependencies required to run the
# Aquaculture Machine Learning Platform. Version numbers are pinned for
# reproducible builds and deployment consistency.
#
# DEPENDENCY CATEGORIES:
# - Core Web Framework: FastAPI and ASGI server
# - Database: PostgreSQL drivers and ORM
# - Caching: Redis clients and extensions
# - Message Processing: Kafka and Celery for async tasks
# - Security: Authentication and encryption libraries
# - Machine Learning: PyTorch, data processing, computer vision
# - Monitoring: Metrics collection and logging
# - Utilities: Configuration, HTTP clients, file handling
#
# MAINTENANCE:
# - Review and update versions quarterly
# - Test compatibility before version bumps
# - Use `pip-audit` to check for security vulnerabilities
# - Pin versions to prevent unexpected production issues
# ============================================================================

# ============================================================================
# CORE WEB FRAMEWORK
# ============================================================================
# Modern Python web framework with automatic API documentation

fastapi==0.104.1                    # Fast, modern web framework with auto docs
uvicorn[standard]==0.24.0           # ASGI server with performance optimizations
pydantic==2.5.0                     # Data validation using Python type hints
pydantic-settings==2.1.0            # Settings management with validation

# ============================================================================
# DATABASE LAYER
# ============================================================================
# PostgreSQL integration with async support and migrations

sqlalchemy==2.0.23                  # Python SQL toolkit and ORM
alembic==1.12.1                     # Database migration tool for SQLAlchemy
psycopg2-binary==2.9.9              # PostgreSQL adapter (binary distribution)
asyncpg==0.29.0                     # Async PostgreSQL driver for better performance

# ============================================================================
# REDIS & CACHING
# ============================================================================
# High-performance caching and session storage

redis==5.0.1                        # Redis client with async support
hiredis==2.2.3                      # High-performance Redis protocol parser

# ============================================================================
# MESSAGE QUEUE & TASK PROCESSING
# ============================================================================
# Distributed task processing and event streaming

kafka-python==2.0.2                 # Apache Kafka client for event streaming
celery==5.3.4                       # Distributed task queue for background jobs
kombu==5.3.4                        # Messaging library (Celery dependency)

# ============================================================================
# AUTHENTICATION & SECURITY
# ============================================================================
# JWT tokens, password hashing, and secure authentication

python-jose[cryptography]==3.3.0    # JWT token handling with crypto support
passlib[bcrypt]==1.7.4              # Password hashing library
python-multipart==0.0.6             # Multipart form data parsing
bcrypt==4.1.1                       # Strong password hashing algorithm

# ============================================================================
# MACHINE LEARNING & DATA SCIENCE
# ============================================================================
# Deep learning, computer vision, and data processing libraries

torch==2.1.1                        # PyTorch deep learning framework
torchvision==0.16.1                 # Computer vision models and transforms
numpy==1.24.3                       # Numerical computing foundation
pandas==2.1.3                       # Data manipulation and analysis
scikit-learn==1.3.2                 # Machine learning algorithms and tools
Pillow==10.1.0                      # Image processing library
opencv-python==4.8.1.78             # Computer vision and image processing
albumentations==1.3.1               # Image augmentation for ML training

# ============================================================================
# MONITORING & LOGGING
# ============================================================================
# Metrics collection, monitoring, and structured logging

prometheus-client==0.19.0           # Prometheus metrics client library
prometheus-fastapi-instrumentator==6.1.0  # FastAPI Prometheus integration
python-json-logger==2.0.7           # Structured JSON logging

# ============================================================================
# HTTP & ASYNC UTILITIES
# ============================================================================
# HTTP clients and asynchronous file operations

httpx==0.25.2                       # Modern async HTTP client library
aiohttp==3.9.1                      # Async HTTP client/server framework
aiofiles==23.2.1                    # Async file operations

# ============================================================================
# DEVELOPMENT & CONFIGURATION UTILITIES
# ============================================================================
# Environment configuration and CLI tools

python-dotenv==1.0.0                # Environment variable management
click==8.1.7                        # Command-line interface creation toolkit
psutil==5.9.6
pyyaml==6.0.1
tenacity==8.2.3

# Data Validation
email-validator==2.1.0